moved buttoncontent to button.jsx
